Rename boto_extra to libboto
